Fixing Germ instantiation error, tissue vs collider and some issues related to drag and drop
The way I did it is as follows, i created an additional script for the coroutine of germ attack and attached it to germ object in the scene. Hence, I erased that part from GermAnimation and passed other values that need to be Serialized in the germ prefab through the germfactory script after instantiating the germ. Afterwards, I realized that there is an issue with the drag so i added a bool value to check if i can drag the object which was the solution.